clib: respect exceptional control flow in `cinit` variable lifting
`cinit` and related methods are able to automatically abstract accesses
of Simpl state variables to Isabelle variables, provided they can prove
that the Simpl variable has not been modified up to the point it is
accessed. However, previously, the automation was unaware of exceptional
control flow. This limits the effectiveness of variable lifting in
situations like the following:

    // `var` has not yet been modified.
    if (condition) {
        var = new_value;
        // Here, `var` has been modified.
        return;
    }
    // Has `var` been modified before the following access?
    do_something(var);

Prior to this commit, the answer would be "yes": `cinit` would conclude
that `var` has been modified prior to the access for `do_something`, so
the variable access would not be abstracted.

With this commit, the answer is "no": `cinit` recognises the `return` in
the `if` block, and can abstract the variable access for `do_something`.

The new automation is enabled for `cinit`, `clift` and `ctac`. It is
currently disabled for `csymbr`, since the new behaviour breaks some
existing proofs.

Signed-off-by: Matthew Brecknell <Matthew.Brecknell@data61.csiro.au>